Creating station and topology factories.
